Keep a reference to only one collection instance
Also fix a nasty javascript aliasing issue that could bite me later. The
cache on CigarFinderWeb.Collections.CigarSearchResults was persisting
after navigating back to the index, even though it was creating a new
instance. While this is actually the behavior I'd like, it was actually
due to a bug.

cachedResponses was actually a property of the prototype, not of the
individual instance, and doing this.cachedResponses[cigar_name] =
something was actually mutating the object on the prototype. It's been
moved to an instance method, and the collection was changed to a
singleton.

This also fixes the funky bug that was occuring when I navigated to and
from the index too quickly. It appears that spamming the geolocation
method was causing it to fail. This should probably be moved to its own
class and really ensured it's only run once, even if the loadLocation
method is called multiple times before it finishes loading.