Change how the server is booted
We're intermittently getting 500s from an invalid constraint error. The
only thing I can think of that could be causing this is that our
requests are getting logged twice. If they are actually getting run
twice, we could end up with a race condition. It looks like it could be
an issue with the `rails server` command.

https://github.com/heroku/rails_12factor/issues/10